handled error

+++ src/Modal/Services/apiCabins.js
@@ -23,28 +23,37 @@ export async function createCabin(newCabin){
     .insert([
     {...newCabin, image:imagePath }
     ])
-
     if(error){
         console.error(error)
         throw new Error ('Cabin could not be Created')
     }
 
-    // File for Image Upload
+    //Image Upload
     const {  error:storageError } = await supabase
     .storage
     .from('cabin-images')
     .upload(imageName,newCabin.image)
+
+    if(storageError){
+        await supabase
+        .from('cabins')
+        .delete()
+        .eq('id', data.id)
+        throw new Error ("Cabin image could not be uploaded and Cabin could not  be created")
+    }
+        return data
+   
+    }
+    // Delete the cabin if there was an error regarding uploading the corresponding image
+
+    export async function deleteCabin(id){
+        const {data, error } = await supabase
+        .from('cabins')
+        .delete()
+        .eq('id', id)
+        if(error){
+            console.error(error)
+            throw new Error ('Cabin could not be deleted')
+        }
         return data
     }
-
-export async function deleteCabin(id){
-    const {data, error } = await supabase
-    .from('cabins')
-    .delete()
-    .eq('id', id)
-    if(error){
-        console.error(error)
-        throw new Error ('Cabin could not be deleted')
-    }
-    return data
-}
